fix: never render HeaderBar without runtime provider
