+++ adapter/src/components/styles/AppWrapper.style.js
